+++ src/buffer.hpp
@@ -21,7 +21,7 @@
  *
  *    Description:  
  *
- *        Version:  0.2
+ *        Version:  0.3.0.1
  *        Created:  28/12/2013 07:01:58
  *       Revision:  none
  *       Compiler:  gcc