+++ src/buffermanager.hpp
@@ -21,7 +21,7 @@
  *
  *    Description:  
  *
- *        Version:  0.2
+ *        Version:  0.3.0.1
  *        Created:  30/12/2013 12:07:43
  *       Revision:  none
  *       Compiler:  gcc