+++ src/gloseq.cpp
@@ -21,7 +21,7 @@
  *
  *    Description:  
  *
- *        Version:  0.2
+ *        Version:  0.3.0.1
  *        Created:  27/12/2013 10:38:49
  *       Revision:  none
  *       Compiler:  gcc