+++ src/rules.cpp
@@ -21,7 +21,7 @@
  *
  *    Description:  
  *
- *        Version:  0.2
+ *        Version:  0.3.0.1
  *        Created:  31/12/2013 18:33:14
  *       Revision:  none
  *       Compiler:  gcc