+++ src/ruleserror.hpp
@@ -21,7 +21,7 @@
  *
  *    Description:  
  *
- *        Version:  0.2
+ *        Version:  0.3.0.1
  *        Created:  06/01/2014 10:04:32
  *       Revision:  none
  *       Compiler:  gcc