+++ src/serialprop.cpp
@@ -21,7 +21,7 @@
  *
  *    Description:  
  *
- *        Version:  0.2
+ *        Version:  0.3.0.1
  *        Created:  21/12/2013 01:39:53
  *       Revision:  none
  *       Compiler:  gcc