+++ src/serialprop.hpp
@@ -21,7 +21,7 @@
  *
  *    Description:  
  *
- *        Version:  0.2
+ *        Version:  0.3.0.1
  *        Created:  19/12/2013 23:25:14
  *       Revision:  none
  *       Compiler:  gcc