+++ src/ultimate.cpp
@@ -21,7 +21,7 @@
  *
  *    Description:  Interface for Ultimate props
  *
- *        Version:  0.2
+ *        Version:  0.3.0.1
  *        Created:  19/12/2013 20:28:22
  *       Revision:  none
  *       Compiler:  gcc